multiboard.j: fix descriptions of color values, its modulo
Previously I incorrectly described the r,g,b,a color values as if they
were clamped to (0,255). This is not the case. The input color value is
actually normalized with "value mod 256" or something similar, like an
integer division.

The result is that you get the magenta color for (-257, -256, -257) as
if you entered (255,0,255). The same applies to TimerDialogSetTitleColor
and probably other APIs of this kind.

Relatably, the float angles of unit facing are also normalized to be in
360 range.

Test code:

	mb = CreateMultiboard()
	-- yes, it's mod
	MultiboardSetTitleTextColor(mb, -257, -256, -257, 255)
	MultiboardSetTitleText(mb, "mboard title")
	MultiboardSetRowCount(mb, 1)
	MultiboardSetColumnCount(mb, 1)
	MultiboardSetItemsValue(mb, "ivalue")
	-- yes it's mod too
	MultiboardSetItemsValueColor(mb, -257, -256, -257, 255)
	MultiboardDisplay(mb, true)