Add notes to Point and Build order functions
Please re-check and reflect on the interpretation.

I marked `BlzQueueBuildOrderById` returning `true` as a note instead of a bug because enqueuing build orders without having the resources for the structure is actually allowed. But it should not be allowed as the first order, so it's debatable if `BlzQueueBuildOrderById` should return `false` in that case if the queue was empty.